feat: disable webln create invoice button on invalid amount
closes #380
