+++ src/style/index.scss
@@ -180,6 +180,7 @@ hr.spacer {
 
 .btn:disabled:hover,
 .btn:disabled {
+    color: #14283f;
     background-color: #545c65;
     cursor: not-allowed;
 }
